+++ CHANGELOG.md
@@ -1,9 +1,14 @@
-# WIP
+# 1.3.2 - Mar 30, 2025
+
+`clojure+.test`:
 
 - Improved capture-output
 - Improved reporting for `(is (not`, `(is (=` and `(is (not=` for 3+ arguments
 - Show `clojure.data/diff` for `(is (=` and collections
-- `clojure+.core/print-class-tree` #6
+
+`clojure+.core`:
+
+- `print-class-tree` #6
 
 # 1.3.1 - Mar 21, 2025